fix robustness testing

+++ mmdet/datasets/pipelines/loading.py
@@ -14,7 +14,7 @@ class LoadImageFromFile(object):
     def __call__(self, results):
         img = mmcv.imread(
             osp.join(results['img_prefix'], results['img_info']['filename']))
-        results['img'] = img.astype(np.float32)
+        results['img'] = img
         results['ori_shape'] = img.shape
         return results
